completed

+++ script.js
@@ -32,6 +32,7 @@ addBox.addEventListener("click", () => {
 });
 
 closeIcon.addEventListener("click", () => {
+  isUpdate = false;
   titleTag.value = "";
   descTag.value = "";
   addBtn.innerText = "Add Note";
@@ -75,6 +76,8 @@ function showMenu(elem) {
 }
 
 function deleteNote(noteId) {
+  let confirmDel = confirm("Are you sure you want to delete this note?");
+  if (!confirmDel) return;
   notes.splice(noteId, 1); // * removing selected note from array/tasks
   // * saving updated notes to localstorage
   localStorage.setItem("notes", JSON.stringify(notes));
@@ -113,7 +116,8 @@ addBtn.addEventListener("click", (e) => {
     if (!isUpdate) {
       notes.push(noteInfo); // * adding new note to notes
     } else {
-      notes[updateId];
+      isUpdate = false;
+      notes[updateId] = noteInfo; // * updating specified note
     }
     // * saving notes to localstorage
     localStorage.setItem("notes", JSON.stringify(notes));
@@ -121,3 +125,5 @@ addBtn.addEventListener("click", (e) => {
     showNotes();
   }
 });
+
+window.onload(showNotes());
